requirements.txt
streamlit
zhipuai
Pillow
pandas
